move Alfven_speed tests to test_speeds.py
